+++ engine/runtime-compiler/src/main/scala/org/enso/compiler/Passes.scala
@@ -44,6 +44,7 @@ class Passes(config: CompilerConfig) {
       SectionsToBinOp.INSTANCE,
       OperatorToFunction,
       LambdaShorthandToLambda,
+      IfThenElseToCase.INSTANCE,
       ImportSymbolAnalysis,
       AmbiguousImportsAnalysis
     ) ++ (if (config.privateCheckEnabled) {